+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 
 ## 0.6.35
 
+### Added
+
+- Fixed #865: Added formatted text editor with basic toolbar and keyboard shortcuts.
+
 ### Fixed
 
 - Fixed #866: Inconsistent exit behavior when leaving output full screen mode.